requirements.txt
h5py==3.3.0
matplotlib==3.4.3
numpy==1.20.3
Pillow==9.2.0
scikit_learn==1.1.1
tensorboardX==2.5.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.62.3
